Units can now be commanded independently.
