+++ Assets/Scenes/Main.unity
@@ -337,8 +337,8 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 437959116}
-  m_LocalRotation: {x: 0.08306907, y: 0, z: 0, w: 0.9965438}
-  m_LocalPosition: {x: 0, y: 14.5, z: -50}
+  m_LocalRotation: {x: 0.45056537, y: 0.000000035731468, z: -0.000000015930091, w: 0.8927434}
+  m_LocalPosition: {x: 0, y: 40, z: -35}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
@@ -839,8 +839,9 @@ GameObject:
   m_Component:
   - component: {fileID: 1237992594}
   - component: {fileID: 1237992593}
+  - component: {fileID: 1237992595}
   m_Layer: 0
-  m_Name: Unit Selection Handler
+  m_Name: Unit Handlers
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -860,7 +861,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _layerMask:
     serializedVersion: 2
-    m_Bits: 0
+    m_Bits: 1
 --- !u!4 &1237992594
 Transform:
   m_ObjectHideFlags: 0
@@ -875,6 +876,22 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1237992595
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1237992592}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c3795f850a59a4f13a819f5fe08d106c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _unitSelectionHandler: {fileID: 1237992593}
+  _layerMask:
+    serializedVersion: 2
+    m_Bits: 1
 --- !u!1 &1356001563
 GameObject:
   m_ObjectHideFlags: 0
