+++ Assets/Scripts/Units/Unit.cs
@@ -6,8 +6,11 @@ using UnityEngine.Events;
 
 public class Unit : NetworkBehaviour
 {
+    [SerializeField] private UnitMovement _unitMovement = null;
     [SerializeField] private UnityEvent onSelected = null;
     [SerializeField] private UnityEvent onDeselected = null;
+
+    public UnitMovement GetUnitMovement() { return _unitMovement; }
 
     #region Client
     public void Select()
