+++ Assets/Scripts/Units/UnitCommandGiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3795f850a59a4f13a819f5fe08d106c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
